calculate speed read from can, to test

+++ Drivers/UserDrivers/DWINTFT/DWINTFT_Radar.h
@@ -13,7 +13,7 @@ void TFT_SetRadarOrder(UART_HandleTypeDef *huart, uint8_t *pRadarOrder, uint8_t 
 void TFT_ExchangeRadarOrder(UART_HandleTypeDef *huart, uint8_t n1, uint8_t n2, uint8_t MaxProbeNum);
 void TFT_DispRadarDist(UART_HandleTypeDef *huart, uint8_t *pRadarDist, uint8_t n);
 void TFT_DispRadarColor(UART_HandleTypeDef *huart, uint8_t *pRadarColor, uint8_t MaxProbeNum);
-void TFT_DispVehicleSpeed(UART_HandleTypeDef *huart, uint8_t speed);
+void TFT_DispVehicleSpeed(UART_HandleTypeDef *huart, uint16_t speed);
 void TFT_ReadProbeOrder(UART_HandleTypeDef *huart);
 
 extern uint8_t RadarProbeOrder[];
